requirements.txt
ultralytics>=8.0.145
streamlit>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.3
matplotlib>=3.7.1
plotly>=5.15.0
pandas>=2.0.3 
numpy>=1.24.3
streamlit-authenticator>=0.2.2
bcrypt>=4.0.1
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=1.0.3
psutil>=5.9.5
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
Pillow>=10.0.0
seaborn>=0.12.2
torch>=2.0.1